requirements.txt
sentence-transformers
fastapi
uvicorn
einops
torch
